+++ HowManySet/Presentation/Common/NavigatorableTextField/NavigatableTextField.swift
@@ -95,28 +95,3 @@ final class NavigatableTextField: UITextField {
     }
     
 }
-
-final class TextFieldNavigator {
-    
-    private var textfields: [NavigatableTextField] = []
-    
-    func register(_ textfields: [NavigatableTextField]) {
-        self.textfields = textfields
-        self.textfields.forEach { $0.navigator = self }
-    }
-    
-    func didTappedPreviousButton(from textfield: NavigatableTextField) {
-        guard let index = textfields.firstIndex(of: textfield),
-              index > 0 else { return }
-        let previousTextField = textfields[index - 1]
-        
-        previousTextField.becomeFirstResponder()
-    }
-    
-    func didTappedNextButton(from textfield: NavigatableTextField) {
-        guard let index = textfields.firstIndex(of: textfield),
-              index < textfields.count - 1 else { return }
-        let nextTextField = textfields[index + 1]
-        nextTextField.becomeFirstResponder()
-    }
-}
